requirements_calc.txt
numpy ; platform.machine == 'arm64'
matplotlib ; platform.machine == 'arm64'
Pillow ; platform.machine == 'arm64'


